requirements.txt
# requirements.txt
torch
torchvision
requests
python-dotenv
tqdm
Pillow